requirements.txt
arrow==0.13.2
astroid==2.3.2
certifi==2019.6.16
configparser==3.5.0
coverage==4.5.1
Django==2.1.7
django-ajax-selects==1.7.1
django-bootstrap-modal-forms==1.4.4
django-chosen==0.1
django-crispy-forms==1.7.2
django-cursor-pagination==0.1.4
django-datatables-view==1.19.1
django-datetimepicker==3.14
django-filter==2.2.0
django-mathfilters==0.4.0
django-pure-pagination==0.3.0
django-simple-history==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
docopt==0.6.2
et-xmlfile==1.0.1
html5lib==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
num2words==0.5.9
numpy==1.15.4
openpyxl==2.6.2
p==0.0.0
pandas==0.23.4
Pillow==5.4.1
pipenv==2018.11.26
postgres==2.2.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
PyPDF2==1.26.0
python-dateutil==2.7.5
pytz==2018.9
reportlab==3.5.13
six==1.12.0
South==1.0.2
typed-ast==1.4.0
virtualenv==16.2.0
virtualenv-clone==0.5.3
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
